refactor(models): fold SGLang server helpers into models.py, drop utils/sglang.py
check_server_health + get_model_id are sync httpx helpers for the sglang setup
path (build_model_factory); their only consumer is core/models.py, so move them
next to sglang_model_factory. First step toward dissolving the utils/ layer.
